+++ dataiter/__init__.py
@@ -94,6 +94,7 @@ from dataiter.data_frame import DataFrameColumn # noqa
 from dataiter.geojson import GeoJSON # noqa
 from dataiter.list_of_dicts import ListOfDicts # noqa
 from dataiter import dt # noqa
+from dataiter import regex # noqa
 
 from dataiter.aggregate import all # noqa
 from dataiter.aggregate import any # noqa