+++ doc/output.py
@@ -11,6 +11,7 @@ sys.path.insert(0, str(Path(".")))
 import dataiter as di
 import numpy as np
 from dataiter import dt
+from dataiter import regex
 di.PRINT_MAX_ITEMS = 3
 di.PRINT_MAX_ROWS = 10
 di.PRINT_MAX_WIDTH = 72